Fix MeasureButton
After changing Measure, it was not checking itself for meter or foot.

+++ measurebutton.h
@@ -35,7 +35,7 @@ public slots:
 signals:
 protected:
 private:
-  int myUnit;
+  int64_t myUnit;
   int myFoot;
 };
 #endif
